pyproject.toml
[project]
name = "enveloppe-mkdocs"
version = "0.0.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "mkdocs",
    "mkdocs-material",
    "mkdocs-awesome-pages-plugin",
    "mdx_breakless_lists",
    "mkdocs-embed-file-plugin",
    "mkdocs_custom_fences",
    "mkdocs-git-revision-date-localized-plugin",
    "mkdocs-encryptcontent-plugin",
    "mkdocs-callouts",
    "mkdocs-custom-tags-attributes",
    "mkdocs-meta-descriptions-plugin",
    "mkdocs-glightbox",
    "mkdocs-obsidian-links",
    "beautifulsoup4",
    "babel"
]

[tool.ruff.lint]
select = ["PTH", "ANN", "N", "Q", "PL", "E", "F", "I"]
ignore = ["E501"]
exclude = ["tests", "docs", "build", "dist", "venv", "venv3", "venv3.6", "venv3.7", "venv3.8", "venv3.9", "venv3.10", "__pycache__"]

[tool.ruff.lint.flake8-quotes]
inline-quotes = "double"

[dependency-groups]
dev = [
    "gitpython>=3.1.43",
    "pydantic>=2.10.4",
    "pyyaml>=6.0.2",
    "rich>=13.9.4",
]
